Question component structure change.
