+++ src/app/app.module.ts
@@ -25,9 +25,10 @@ import { NgApexchartsModule } from 'ng-apexcharts';
 import { BarChartsComponent } from './components/analytics/charts/barcharts.component';
 import { PiechartsComponent } from './components/analytics/charts/piecharts.component';
 import { MatInputModule } from '@angular/material/input';
-import { RadioQuestionComponent } from './components/submission/radio-question.component';
-import { MultipleChoiceQuestionComponent } from './components/submission/multiple-choice-question.component';
-import { ShortAnswerQuestionComponent } from './components/submission/short-answer-question.component';
+import { RadioQuestionComponent } from './components/submission/questions/question-radio.component';
+import { MultipleChoiceQuestionComponent } from './components/submission/questions/question-multiple-choice.component';
+import { ShortAnswerQuestionComponent } from './components/submission/questions/question-short-answer.component';
+import { DropdownQuestionComponent } from './components/submission/questions/question-dropdown.component';
 
 
 @NgModule({
@@ -46,7 +47,8 @@ import { ShortAnswerQuestionComponent } from './components/submission/short-answ
     PiechartsComponent,
     RadioQuestionComponent,
     MultipleChoiceQuestionComponent,
-    ShortAnswerQuestionComponent
+    ShortAnswerQuestionComponent,
+    DropdownQuestionComponent
   ],
   imports: [
     BrowserModule,
