+++ src/app/components/submission/questions/question-short-answer.component.ts
@@ -6,11 +6,11 @@ import { IQuestion } from 'src/app/models/iquestion-question';
   template: `
     <br><hr><br>
     <div class="mb-3">
-      <label for="{{'question-'+index+'-'+'form'}}" class="form-label">{{ question.title }}<span style="color:red" *ngIf="question.isRequired">*</span></label>
+      <label for="{{'question-'+index+'-'+'form'}}" class="form-label">{{ question.title }}<span *ngIf="question.isRequired">*</span></label>
       <textarea class="form-control" id="{{'question-'+index+'-'+'form'}}" rows="3"></textarea>
     </div>
   `,
-  styleUrls: ['./submission.component.css']
+  styleUrls: ['../submission.component.css']
 })
 
 export class ShortAnswerQuestionComponent {
