+++ src/app/components/submission/submission.component.html
@@ -2,16 +2,13 @@
   <div class="container bubble">
 
     <div class="d-flex justify-content-center">
-      <h1 class="display-4">Training Feedback</h1>
+      <h1 class="display-4">{{survey.title}}</h1>
     </div>
     <hr>
     <div class="container">
 
       <br>
-      <p>Thank you for being a part of Revature. We hope you are having as much fun attending as we do organizing.</p>
-      <p>We want to hear your feedback so we can keep improving our logistics and content. Please fill this quick survey out and let us know your thoughts. Please note that you will be using this SAME link EVERY week to give your feedback about training, so you may want to bookmark or pin this URL in your Slack channel.</p>
-      <p>Survey responses are only accessible by the QC team and are collected anonymously unless you fill in your information below.</p>
-      <p>Please note this survey is for the last completed week of training (e.g. if you are currently in Week 4, please fill out the survey for Week 3.)</p>
+      <p>{{survey.description}}</p>
       <span>* Required</span>
     </div>
 
@@ -72,6 +69,7 @@
           <radio-question id="{{'question-'+i}}" [question]="question" [index]="i" *ngIf="question.questionType == 'RADIO'"></radio-question>
           <multiple-choice-question id="{{'question-'+i}}" [question]="question" [index]="i" *ngIf="question.questionType == 'MULTIPLE_CHOICE'"></multiple-choice-question>
           <short-answer-question id="{{'question-'+i}}" [question]="question" [index]="i" *ngIf="question.questionType == 'SHORT_ANSWER'"></short-answer-question>
+          <dropdown-question id="{{'question-'+i}}" [question]="question" [index]="i" *ngIf="question.questionType == 'DROPDOWN'"></dropdown-question>
 
         </div>
